+++ src/ezdxf/math/_ctypes.py
@@ -1,10 +1,13 @@
-#  Copyright (c) 2020-2021, Manfred Moitzi
+#  Copyright (c) 2020-2022, Manfred Moitzi
 #  License: MIT License
+from typing import Union, Sequence
 from ezdxf.acc import USE_C_EXT
 
 __all__ = [
     "Vec3",
     "Vec2",
+    "AnyVec",
+    "Vertex",
     "X_AXIS",
     "Y_AXIS",
     "Z_AXIS",
@@ -80,3 +83,7 @@ else:
         arc_angle_span_deg,
         arc_angle_span_rad,
     )
+
+# Early required type compositions
+AnyVec = Union[Vec2, Vec3]
+Vertex = Union[Sequence[float], AnyVec]